Fix data registration
